chore: add temporary workaround for dbgpt's aiohttp version requirement (#205)
The PR is a temporary solution. The reason for the issue is that dbgpt restricts aiohttp to version 3.8.4, which is too low, and has a dependency conflict with litellm. We are in contact with dbgpt, and they will address this issue in the next release version of dbgpt. However, to ensure a smooth installation and user experience for Chat2Graph, we are temporarily accepting this "not so clean" code. At the same time, once dbgpt resolves the version dependency issue, we will consider handling your submitted code (possibly removing it) to maintain clean code standards.